+++ src/lib/config/sanity/schemas/settings.ts
@@ -1,11 +1,9 @@
-import { CogIcon } from "@sanity/icons";
 import { defineType } from "sanity";
 
 export default defineType({
   name: "settings",
   title: "Settings",
   type: "document",
-  icon: CogIcon,
   fields: [
     {
       name: "title",
@@ -53,5 +51,6 @@ export default defineType({
   ],
   preview: {
     prepare: () => ({ title: "Settings" }),
+
   },
 });